Fix type of float and complex untyped constant when they assigned to variable of interface type or blank identifier

+++ src/ro/redeul/google/go/lang/psi/impl/expressions/primary/GoLiteralExpressionImpl.java
@@ -82,7 +82,7 @@ public class GoLiteralExpressionImpl extends GoExpressionBase
 
                 case Float:
                     return new GoType[]{
-                            GoTypes.getBuiltin(Builtin.Float32, namesCache)
+                            GoTypes.getBuiltin(Builtin.Float64, namesCache)
                     };
 
                 case Char:
@@ -93,7 +93,7 @@ public class GoLiteralExpressionImpl extends GoExpressionBase
                 case ImaginaryInt:
                 case ImaginaryFloat:
                     return new GoType[]{
-                            GoTypes.getBuiltin(Builtin.Complex64, namesCache)
+                            GoTypes.getBuiltin(Builtin.Complex128, namesCache)
                     };
 
                 case RawString:
